Update injector getter
