+++ scripts/hexo/_inject/injector.js
